Implemented operator options

+++ com.Lockedme/src/com/Lockedme/FileUtil.java
@@ -5,12 +5,20 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.Lockedme.constants.LockedmeConstants;
+import com.Lockedme.models.OperationResponse;
+
 public class FileUtil {
 	static int count = 0;
 	List<FileDetail> files = new ArrayList<>();
+	private String rootPath;
+
+	public FileUtil(String _rootPath) {
+		this.rootPath = _rootPath;
+	}
 
 	public List<FileDetail> getFiles() {
-		File f = new File("F:\\FileIO\\Lockedme");
+		File f = new File(this.rootPath);
 		if (this.files != null) {
 			this.files.clear();
 		} else {
@@ -18,6 +26,9 @@ public class FileUtil {
 		}
 		count = 0;
 		String[] fs = f.list();
+		if (fs == null) {
+			return null;
+		}
 		for (String file : fs) {
 			count++;
 			FileDetail fetchedFile = new FileDetail(count, file);
@@ -27,35 +38,68 @@ public class FileUtil {
 	}
 
 	public void addFile(String fileName) throws IOException {
-		File f = new File("F:\\FileIO\\Lockedme.com\\" + fileName + ".txt");
+		File f = new File(this.rootPath + fileName + ".txt");
 		f.createNewFile();
 	}
 
-	public boolean deletFile(String fileName) {
+	public OperationResponse deletFile(String fileName) {
 		boolean isDeleted = false;
+		OperationResponse result = new OperationResponse();
 		files = getFiles();
-		for (FileDetail file : files) {
-			if (file.getFileName().startsWith(fileName)) {
-				File f = new File("F:\\FileIO\\Lockedme.com\\" + file.getFileName());
-				f.delete();
-				isDeleted = true;
-				break;
+		if (files == null) {
+			result.setSuccess(false);
+			result.setMessage(LockedmeConstants.PATH_NOTEXIST_MESSAGE);
+		} else if (files.isEmpty()) {
+			result.setSuccess(false);
+			result.setMessage(LockedmeConstants.FOLDER_EMPTY_MESSAGE);
+		} else {
+			for (FileDetail file : files) {
+				if (file.getFileName().startsWith(fileName)) {
+					File f = new File(this.rootPath + file.getFileName());
+					f.delete();
+					isDeleted = true;
+					break;
+				}
+			}
+
+			result.setSuccess(isDeleted);
+			if (isDeleted) {
+				result.setMessage(LockedmeConstants.DELETE_MESSAGE);
+			} else {
+				result.setMessage(LockedmeConstants.FILE_NOTEXISTED_MESSAGE);
 			}
 		}
-		return isDeleted;
+
+		return result;
 	}
 
-	public boolean search(String filName) {
+	public OperationResponse search(String filName) {
 		boolean isIdentified = false;
+		OperationResponse result = new OperationResponse();
 		files = getFiles();
-		for (FileDetail file : files) {
-			if (file.getFileName().startsWith(filName)) {
-				isIdentified = true;
-				break;
+		if (files == null) {
+			result.setSuccess(false);
+			result.setMessage(LockedmeConstants.PATH_NOTEXIST_MESSAGE);
+		} else if (files.isEmpty()) {
+			result.setSuccess(false);
+			result.setMessage(LockedmeConstants.FOLDER_EMPTY_MESSAGE);
+		} else {
+			for (FileDetail file : files) {
+				if (file.getFileName().startsWith(filName)) {
+					isIdentified = true;
+					break;
+				}
+			}
+
+			result.setSuccess(isIdentified);
+			if (isIdentified) {
+				result.setMessage(LockedmeConstants.FILE_EXISTED_MESSAGE);
+			} else {
+				result.setMessage(LockedmeConstants.FILE_NOTEXISTED_MESSAGE);
 			}
 		}
-		return isIdentified;
+
+		return result;
 	}
-	
-	
+
 }
